[substrate_api/encointer_api] getNextMeetupTime uses now dart only

+++ packages/ew_encointer_utils/test/assignment_test.dart
@@ -16,7 +16,7 @@ class MeetupTimeTestCase {
   final double longitude;
 
   /// Global `MeetupTimeOffset` of the encointer chain.
-  final double offset;
+  final int offset;
 
   /// Expected result in ms units.
   final int expected;
